fix(): add failing tests
