+++ src/test/java/io/kestra/plugin/gcp/bigquery/CopyPartitionsTest.java
@@ -33,6 +33,7 @@ class CopyPartitionsTest {
         Query create = Query.builder()
             .id(QueryTest.class.getSimpleName())
             .type(Query.class.getName())
+            .projectId(Property.ofValue(project))
             .sql(Property.ofValue("CREATE TABLE `" + project + "." + dataset + "." + table + "` (transaction_id INT64, transaction_date DATETIME)\n" +
                     "PARTITION BY DATE(transaction_date)\n" +
                     "AS (SELECT 1, DATETIME '2020-04-01 12:30:00.45')\n" +
@@ -70,6 +71,7 @@ class CopyPartitionsTest {
         Query query = Query.builder()
             .id(QueryTest.class.getSimpleName())
             .type(Query.class.getName())
+            .projectId(Property.ofValue(project))
             .fetchOne(true)
             .sql(Property.ofValue("SELECT COUNT(*) as cnt FROM `" + project + "." + dataset + "." + destinationTable + "`;"))
             .build();
