+++ src/test/java/io/kestra/plugin/gcp/bigquery/LoadTest.java
@@ -53,6 +53,7 @@ class LoadTest {
         Load task = Load.builder()
             .id(LoadTest.class.getSimpleName())
             .type(Load.class.getName())
+            .projectId(Property.ofValue(project))
             .from(Property.ofValue(source.toString()))
             .destinationTable(Property.ofValue(project + "." + dataset + "." + FriendlyId.createFriendlyId()))
             .format(AbstractLoad.Format.CSV)
@@ -85,6 +86,7 @@ class LoadTest {
         Load task = Load.builder()
             .id(LoadTest.class.getSimpleName())
             .type(Load.class.getName())
+            .projectId(Property.ofValue(project))
             .from(Property.ofValue(source.toString()))
             .destinationTable(Property.ofValue(project + "." + dataset + "." + FriendlyId.createFriendlyId()))
             .format(AbstractLoad.Format.AVRO)
@@ -113,6 +115,7 @@ class LoadTest {
         Load.LoadBuilder<?, ?> task = Load.builder()
             .id(LoadTest.class.getSimpleName())
             .type(Load.class.getName())
+            .projectId(Property.ofValue(project))
             .from(Property.ofValue(source.toString()))
             .destinationTable(Property.ofValue(project + "." + dataset + "." + FriendlyId.createFriendlyId()))
             .format(AbstractLoad.Format.CSV)
