+++ src/test/java/io/kestra/plugin/gcp/bigquery/StorageWriteTest.java
@@ -2,6 +2,7 @@ package io.kestra.plugin.gcp.bigquery;
 
 import com.devskiller.friendly_id.FriendlyId;
 import com.google.common.collect.ImmutableMap;
+import io.kestra.core.junit.annotations.KestraTest;
 import io.kestra.core.models.property.Property;
 import io.kestra.core.runners.RunContext;
 import io.kestra.core.runners.RunContextFactory;
@@ -11,7 +12,7 @@ import io.kestra.core.tenant.TenantService;
 import io.kestra.core.utils.IdUtils;
 import io.kestra.core.utils.TestsUtils;
 import io.micronaut.context.annotation.Value;
-import io.kestra.core.junit.annotations.KestraTest;
+import jakarta.inject.Inject;
 import org.junit.jupiter.api.Test;
 
 import java.io.File;
@@ -25,7 +26,6 @@ import java.time.ZonedDateTime;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
-import jakarta.inject.Inject;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
@@ -51,6 +51,7 @@ class StorageWriteTest {
         Query create = Query.builder()
             .id(QueryTest.class.getSimpleName())
             .type(Query.class.getName())
+            .projectId(Property.ofValue(project))
             .sql(Property.ofValue("CREATE TABLE " + table + " AS " + "SELECT \n" +
                 "  \"hello\" as string,\n" +
                 "  CAST(NULL AS INT) AS `nullable`,\n" +
@@ -96,6 +97,7 @@ class StorageWriteTest {
         StorageWrite task = StorageWrite.builder()
             .id("test-unit")
             .type(StorageWrite.class.getName())
+            .projectId(Property.ofValue(project))
             .destinationTable(Property.ofValue(table))
             .location(Property.ofValue("EU"))
             .from(Property.ofValue(put.toString()))
@@ -104,8 +106,6 @@ class StorageWriteTest {
         RunContext runContext = TestsUtils.mockRunContext(runContextFactory, task, ImmutableMap.of());
         StorageWrite.Output run = task.run(runContext);
 
-
-//        assertThat(run.getRowsCount(), is(1));
         assertThat(run.getRows(), is(1));
     }
 }
