--- src/test/java/io/kestra/plugin/gcp/gcs/InferedProjectIdBucketTest.java
@@ -1,8 +1,0 @@
-package io.kestra.plugin.gcp.gcs;
-
-public class InferedProjectIdBucketTest extends AbstractBucketTest {
-    @Override
-    public boolean inferProjectId() {
-        return true;
-    }
-}
