+++ src/test/java/io/kestra/plugin/gcp/pubsub/TriggerTest.java
@@ -75,7 +75,6 @@ class TriggerTest {
                 assertThat(execution.getLeft().getFlowId(), is("pubsub-listen"));
             });
 
-
             worker.run();
             scheduler.run();
